Integrate new features into getTranslationsFromExpression() [CU-860rafnru].
Previous work on internal helpers to manipulate calls made to translate() drastically simplified this function. Its doc was simplified to avoid repetition, its unit tests were also simplified, and the focus is now on the recursion mechanism. It used to be derived from tools::xgettext() but it its current form, it feels much more modern and cleaner. I still refer to my initial inspiration for transparency purposes, but there is not much in common between these functions.