Initial commit on third implementation of the source parser/compiler [WIP].
WARNING! This is still a WIP. I must reorganize and clean new mechanisms, document them, and test them. Many other existing features will be either updated or deprecated following its integration. Existing unit tests must also be updated accordingly.

The entry point is extract(). It extracts and parses text to be translated. Its output can be (almost) passed as is to a future write() mechanism. The latter will complete the Intl compiler of the package.