Introduce big changes everywhere (yet again) [WIP] [CU-8678ran8h].
This commit brings a lot of changes that address design challenges. Notably, stopFormat() and stopSilently() are now deprecated in favour of halt(). Class SrcTranslateCall has become SrcTranslation and completely changed. I also introduced a lot of tweaks everywhere to keep the code consistent with latest changes.

The design can now be summarized by the chain raw expression (as string) -> SrcLoc -> SrcExpr -> isTranslateCall -> SrcTranslation -> SrcString. This chain relies on coercions instead of inheritance. This is a lot better than what I previously had in mind and should likely become part of a more stable API in the future.

There is a bug that remain to be fixed when printing expressions. Turns out "expressions" (at least according to the parser) is a broader category of objects compared to expression objects themselves.

The last method (asSrcString.SrcTranslation()) is the final piece of the puzzle to be implemented to complete the extract() mechanism. It will be done next week.