Initial commit on the assert*() family of functions (WIP) [CU-8678ran8h].
This includes their integration into all existings features. This family dramatically simplifies input checking by standardizing all error messages and encapsulating defensive programming techniques. The slight overhead it introduces is negligible because they are only relevant when there is an error to convey to the user.

Functions assertSingle*InRange() may need a little bit of optimization / further work when generating intelligent error messages passed to halt(). This requires an auxiliary function. I will think about it a bit, but this is nothing fancy.