Initial commit on read_translations_file().
This commit includes all underlying (internal) core mechanisms to read, extract, and parse contents
from such a file. It works fine, and it is fast, but I already included small artifacts to guide my
next steps. These break the code, but it is small and I prefer to keep them for guidance. I will
most likely refactor and reorganize internal mechanisms and work on a R6 TranslatorEnv class as I had
planned a long time ago.