Rename script read-translations.R to translation-io.R and more (WIP).
Furthermore, initial commit on tests and doc of read_translations(). This includes mocks and
placeholders for write_translations().

Began work on centralizing documentation on translations source files into doc of *_translations()
functions because it is super relevant to users. Therefore, doc of internal mechanisms is now a bit
more concise to avoid duplication.

Obviously, code and doc will be properly updated once class Translator will be updated. There is
a couple of placeholders.

This is still a WIP. I need to integrate a code chunk that outputs doc taken from READMEs into the
RD doc. R CMD check is also throwing a lot of errors.