Revamp of S3 class Block.
The latter has now more elements and is a bit more complete. This commit also includes doc and tests.
However, it breaks parsing mechanisms of source blocks in TSFs. This will be fixed in a future commit.
In the meantime, I have removed underlying unit tests of read_translations() to avoid throwing (known)
errors.