Initial commit on refactoring of all TSF parsing mechanisms (WIP).
This is still a work in progress. I am in the process of merging all mechanisms in the same script. I
renamed a couple of functions implemented before for consistency, and some mechanisms are being replaced.
I have to modify some existing tests, write new ones, and centralize internal doc on parsing. See next
commit.