Major revamp of class Location (WIP).
The class is now much faster. It was not refactoresed as an R6 class because
the overhead is simply not worth it. This commit further introduces new S3
methods: a revamped format() and a new c().

There is also a new merge_locations() function. c() can only merge ranges
of Location objects having the same path. merge_locations() generalizes this
idea to multiple paths (it calls c() for groups of Location objects).