Major revamp of class Block (WIP).
The class is now much faster. It was refactored as an R6 class because
the overhead is worth it. I really needed to hide away the underlying
the data behind a robust and consistent API.

I also decided to defer many features directly to this class. It is now
in charge of hashing translations. I further decided sha1 would be the
only hashing algorithm used by transltr. This simplifies the UX. It will
have an impact on TSFs later on (nothing big).

I have used tricks in my sleeves to make it as fast and light as possible.

This commit further introduces new S3 generics and methods: a revamped
format(), a new c(), a new as_block() S3 generic, and a as_block.call()
method.

There is also a new merge_blocks() function that is very similar to
merge_locations(). c() can only merge Block objects having the same
hash (same source key AND text). merge_blocks() generalizes this idea to
multiple hashes (it calls c() on groups of Block objects).