Revamped translate() interface (WIP).
This commit also introduces S3 generics is_translate_call() and
match_translate_call() to identify and process calls to transltr::translate().

While is_translate_call() is fast enough, I am well aware that relying on
regex to parse R code is not the best. For now, it works fine and it is
fast enough. However, in the future, I would really like to write an
optimized tokenizer/parser that could further handle R code such as
call("translate") and do.call(translate, list(...)). This is not important
for now.