Initial commit on find_translations() (WIP).
This is a very important mechanism and pretty much the heart of the package.
This function (and its lower-level helpers) is in charge of parsing all R
scripts within a project, detect translations, construct Block objects for
them and intelligently merge them. It uses features introduced in earlier
commits.

The final objective is to output a Translator object. This will come later in
a future PR.