Robustify block() and .block().
They now perform some input validation, instead of deferring the whole process to lower-level
mechanisms. This is better for safety and also for consistency, because underlying error messages
are more semantic (they use the 'real' argument names). There is a slight penalty in terms of
performance, but I am okay with it.