Final implementation of text_normalize().
This includes related documentation, unit tests, and other side-files. It is ready
to be used.

I will be honest: I really dislike the current implementation, and I wrote a @note
on this in the documentation. I dislike how it is written, I dislike how I depend
on 5 calls to gsub(), and I am not satified with current performance. Calling
text_normalize() on empty dots takes ~18 micro-seconds on my machine. Typical calls
takes about 24 to 36 micro-seconds. However, since this function is called everytime
translate() is called, I am shooting for something that can consistently run in less
than 15 micro-seconds. 10 micro-seconds would be ideal, but I doubt it is actually
possible in pure R, and I do not want to introduce C/C++ compiled code at runtime.