New implementation of the translator_*() family (WIP).
This re-introduces functions translator_set(), translator_get(), and introduces new internal
functions translator_scope() and translator_scope_name(). They now uses an internal cache stored
in the package (it is not attached to the search path). This is still a WIP because some unit tests
are missing. Documentation is also missing.

This implementation is much more robust. It is not as fast, but the extra resources balances a much
more convenient approach that is closer to what user expects. As such, it would be much harder to
break.