Bugfix: do not use sys.parents() in translator_scope().
sys.parents() is not necessarily a strictly increasing sequence, and may not match seq_len(sys.nframe()).
This is not what we want in the current implementation. Having said that, there could be a way to
exploit sys.parents() to traverse the call stack more efficiently. This could be revisited in the future.