Fix: Kana, and non-ASCII characters cannot be used in Rd file.
They were removed, and replaced by ASCII chararacters that appears to be
supported (at least on my machine?).

Turns out that even though R natively supports UTF-8 now, the underlying Rd
does not (very poor support). This is because it relies on LaTeX (and pandoc)
to convert Rd files into PDF manuals. And LaTeX, as of 2024, still has poor
UTF-8 support. The chain appears to be

R -> Rd2pdf -> pandoc -> LaTeX -> pdflatex.

This is problematic, and I am not afraid to say that the R community should be
ashamed of that. People cannot even build manuals in their preferred language.
Relying on a huge chain of programs and transitive dependencies is not always
a wise idea, and we are now paying the price.

I would advise R to switch to HTML entirely.