Refactor portable() mechanisms into export(), and import().
Introduction
------------
While writing tests and documentation for portable(), related classes, and
related methods, it became clear that the current design was not optimal, and
not mature enough. Documentation was lengthy, and held many small details that
hinted towards a bad design. The same can be said for unit tests. While
writing tests for format(), it became apparent that the whole thing was
lacking a clearer direction for design.

Hence, all previous functions were replaced, and revamped. The current
implementation is much more succint, and is much easier to read, document,
and test.

Design
------
Old mechanisms are explained differently. The process of exporting data to
YAML is now called "data serialization" (as expected). It is broken down in
two steps, as explained in the new doc: preprocessing via export(), and
serialization via format().

- export() replaces all portable_*() functions. It emits objects of class
  Exported that are trivially serializable via format.
- import() replaces previous methods as_translator.PortableTranslation(),
  and etc. It transforms Exported objects back into regular R objects.

Much more details are given in the documentation.

Changes
-------

Aside from obvious changes in the old scripts related to portable(), these
big changes had a significant impact on I/O functions for PTF files. There
is still some work to do on these to fully integrate these changes (more to
come very soon).