Initial commit on the Flat format (WIP).
This is still a WIP because it requires further testing. Initial manual
testing shows that this new feature is ready to be formally tested. The
objective is to use this format instead of YAML for Portable Translations
Files. This would make the file even easier to read by non-technical
collaborators (and it would be hard to simplify it any further).

The Flat format is a minimal data serialization format, where R objects
are converted to character strings and organized into unindented (flat)
sections identified by a tag.

Flat strings can be parsed back into R objects in a way that natively
preserves the shape/structure of the original object.

It has an S3 class and a print() method for convenience.