Deprecate scoping mechanisms for Translator objects.
This saddens me but even though it was flagged as being experimental (and
had strong doc and tests), the feature was simply too fragile for non-expert
use. It is much easier to simply pass a Translator argument to translate()
until further notice.

I am still experimeting with ways to register Translator objects (safely) to
avoid passing them as arguments to translate() all the time. This feature is
likely gonna come out with a future version before 1.0.0.