New (revamped) implementation of format_vector() (WIP).
TODO: rewrite documentation, examples, and unit tests.

This new version is faster, better, and more flexible than the previous
one. It also fixes a number of bugs and inconsistencies found in the previous
one. Notably, it can handle any object of length 1, or 0 (named or unnamed).

It is already implemented and tested for functions that rely on it.