Bugfixes: deactivate tests that relies on Sys.chmod() on Windows.
As explained in its documentation, Sys.chmod() does not really work
on this OS. Since the behavior is well tested on other OSes (Ubuntu,
and macOS), it is not required to write a specific test block for
Windows.

This also fixes a small inconsistency on how R use file separators on
Windows. This would not affect the behavior of the package in normal
circumstances, but a fix is required for the (identical) expectation
to work.