New implementation of normalize().
This is a new version that is (1) much better, (2) faster, and (3)
more readable. It is better aligned with the underlying use cases
of the package and is documented in a more thorough way with a new
dedicated roxygen2 template for text normalization.

It is also written in a way that uses 2 calls to gsub(), which is
better than the previous version that used 5 calls. I would still
like to reduce the logic to a single gsub() call in the future.

This new version removes the need of a 'concat' constant. The latter
will be removed from other functionnalities in the next commit.