Website is now generated in another directory.
This directory is in the parent directory of the package's root
directory. This is all done automatically using pkgdown. The
website is now also pushed to a different repository (transltr-www).
The latter does not matter, it only exists for automation purposes.

I personally disliked storing the website in another branch. It felt
more complicated than using a distinct repo and directory.