Shorten names of environment variables.
- Rename TRANSLTR_LANGUAGE to TRANSLTR_LANG.
- Rename TRANSLTR_SOURCE_LANGUAGE to TRANSLTR_SOURCE_LANG.

This is done in accordance with implicit naming conventions of the
package: 'lang' typically refers to languages codes, and 'language'
to their underlying full names (native or not).

Internal constant .__LGL_DEBUG_FALSE is kept (the only exception).
It is absolutely required for testing purposes of language global
setter functions.