Revamp class Translator (WIP).
Tests and documentation are missing and will follow in a future
commit.

This is a huge change done for three reasons: (1) performance,
(2) simplicity, and (3) convenience. The API of the class is now
much easier to use, and faster. It also leverages the features of
the new SourceText class (superseding class Text). All related
functions were updated accordingly.