Update and rename find_source*() functions to find_source_text*() (WIP).
These functions were updated following changes to classes Translator
and SourceText (superseding class Text). Documentation and tests will
follow in a future commit.

I am planning to add a new 'ignore' argument to find_source_text() to
skip files that should not be scanned in a project (such as renv/
directories). This will follow in a future commit.

- All find_source*() functions were merged together into a single
  source script for consistency with how other scripts are structured.
- Other super small changes in styling.