Initial commit on a minimal error reporter class (WIP).
Class ErrorReporter is an internal utility class that accumulates
error messages, formats them, and optionally, throw these errors
when $report(TRUE) is called. It will be most useful in complex
assert() methods.

It is implemented as a stack that grows from top to bottom. Errors
are $push()-ed to it, and can also be $pop()-ped from it. The stack
is implemented as a list with pre-allocated slots that automatically
$grow()s when its capacity is reached (default size is 25). $report()
returns errors, and may throw them all at once too. It comes with
usual functions: a constructor, an introspector, format(), and
print() methods. The latter will be mostly useful for debugging
purposes.

Documentation and tests will follow in a future commit once it
reaches a stable state.