Super early revamp of serialization mechanisms (WIP).
BEWARE. THIS IS INCOMPLETE WORK THAT BREAKS THE PACKAGE.

This commit is a big mess. A LOT of changes are missing, functionalities
are incomplete, documentation is outdated, and so are tests too. It was
pushed mostly for back-up purposes, and to convey my intent to (future)
users.

I have decided to split serialize(), export(), and import() mechanisms
in separate scripts. assert() methods are grouped with export() methods
as they are closely tied.

More to come in the future.